pyproject.toml
[tool.poetry]
name = "pwdetest\b\b\b"
version = "0.1.0"
description = "test fklas"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.3"
requests = "^2.31.0"
pandas = "^2.1.3"
matplotlib = "^3.8.2"
websocket-client = "^1.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
